libc: libstdc++: Introduce libstdc++_vendor
A lot of blobs still link to libstdc++.so even after 8.1, so allow
devices to build a vendor copy of it.

Adding vendor_available to libstdc++ is not an option after Android 14,
since libstdc++ then gets listed on /vendor/etc/linker.config.pb,
which is duplicated with /system/etc/linker.config.pb and as a result
linkerconfig fails and stops the entire system from booting.

[    5.840677] DEBUG: Abort message: 'duplicate: libstdc++.so is provided by default and system in [vendor]'

Change-Id: Ib6eee48de87100c2fb0dd765d57ee95d3159c4ef
Signed-off-by: Dmitrii <bankersenator@gmail.com>